constant renamed, dir renamed
