+++ calendly-clone/src/common/constants.js
@@ -1,5 +1,5 @@
 export const NAVIGATION_PATHS = {
-    CREATE_EVENT_PAGE: '/create-event',
+    USER_EVENTS_PAGE: '/user-events',
     SIGN_IN: '/sign-in',
     SIGN_UP: '/sign-up',
 }
